refactor(engine): Reorganize dsl module structure
