+++ tests/conftest.py
@@ -44,7 +44,7 @@ def auth_sandbox():
     from tracecat.types.auth import Role
 
     service_role = Role(
-        type="service", user_id="test-tracecat-user", service_id="tracecat-testing"
+        type="service", user_id="test-tracecat-user", service_id="tracecat-runner"
     )
     ctx_role.set(service_role)
     yield
