+++ tracecat/dsl/schedules.py
@@ -15,7 +15,8 @@ from temporalio.client import (
 
 from tracecat import config, identifiers
 from tracecat.contexts import ctx_role
-from tracecat.dsl.common import DSLInput, get_temporal_client
+from tracecat.dsl.client import get_temporal_client
+from tracecat.dsl.common import DSLInput
 from tracecat.dsl.workflow import DSLRunArgs, DSLWorkflow
 from tracecat.types.api import UpdateScheduleParams
